# api.yaml
openapi: 3.0.0
info:
  title: Personal-Trainer-API
  version: 1.0.0
servers:
  - url: /api/v1
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    BaseResponse:
      type: object
      required: [status, message, code]
      properties:
        status:
          type: string
          enum: [success, error]
        message:
          type: string
        code:
          type: string
          description: Machine-readable response code (e.g., OK, BAD_REQUEST, NOT_FOUND)
    SuccessResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            data:
              nullable: true
              description: Any JSON value (object/array/string/number/etc)
              x-go-type: interface{}
            meta:
              nullable: true
              description: Any JSON value (usually object)
              x-go-type: interface{}
    ErrorResponse:
      allOf:
        - $ref: "#/components/schemas/BaseResponse"
        - type: object
          properties:
            errors:
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/FieldError"
    FieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          description: The field that failed validation
        message:
          type: string
          description: Human-readable error message for this field
...
    UpdateProfileRequest:
      type: object
      description: |
        Partial profile update. avatar_url is intentionally NOT on this
        request — avatars are set exclusively via
        POST /users/me/profile/picture (which uploads through MinIO and
        writes the column asynchronously). Putting avatar_url here too
        would let the JSON endpoint clobber a freshly-uploaded picture
        and was previously also forcing profile_complete = false until
        a picture existed.
      properties:
        name:
          type: string
          example: "Ezeaku Dorothy"
        gender:
          type: string
          enum: [male, female, other]
          example: female
        fitness_goals:
          type: array
          items:
            type: string
            enum:
              - lose_weight
              - build_muscle
              - improve_flexibility
              - boost_energy
              - build_healthy_habits
              - build_strength
          minItems: 1
          example: [build_muscle, build_strength]
        fitness_level:
          type: string
          enum: [beginner, intermediate, advanced]
          example: beginner
    UserProfileResponse:
      type: object
      required:
        [
          id,
          email,
          name,
          gender,
          fitness_goals,
          fitness_level,
          avatar_url,
          profile_complete,
        ]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        gender:
          type: string
          nullable: true
        fitness_goals:
          type: array
          items:
            type: string
        fitness_level:
          type: string
          nullable: true
        avatar_url:
          type: string
          nullable: true
        profile_complete:
          type: boolean
...
security:
  - bearerAuth: []
paths:
...
  /users/me/profile:
    patch:
      operationId: UpdateUserProfile
      tags:
        - User
      summary: Update the authenticated user's profile (onboarding)
      description: |
        Saves name, gender, fitness goals, and fitness level for the logged-in
        user. All fields are optional — send only what changed. avatar_url is
        NOT accepted here; upload via POST /users/me/profile/picture.
        profile_complete flips to true once name, gender, and fitness_level
        are all set.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfileResponse"
              example:
                status: success
                message: Profile updated successfully
                code: OK
                data:
                  id: "550e8400-e29b-41d4-a716-446655440000"
                  email: "dorothy@example.com"
                  name: "Ezeaku Dorothy"
                  gender: "female"
                  fitness_goals: ["build_muscle", "build_strength"]
                  fitness_level: "beginner"
                  avatar_url: null
                  profile_complete: true
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Missing or invalid access token
                code: UNAUTHORIZED
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    get:
      operationId: GetUserProfile
      tags:
        - User
      summary: Get the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Profile fetched
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UserProfileResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/profile/picture:
    post:
      operationId: UploadProfilePicture
      tags:
        - User
      summary: Upload an avatar image
      description: |
        Upload a profile picture for the authenticated user. The request must be
        `multipart/form-data` with a single `picture` part containing the image
        bytes. Accepted formats: JPEG, PNG, WebP, HEIC.

        Files up to 5 MiB are stored as uploaded. Files over 5 MiB are
        re-encoded as JPEG with progressively lower quality until they fit
        under 5 MiB (final format will be JPEG regardless of input).

        The actual upload to object storage happens **asynchronously** in a
        background worker — this endpoint returns 202 immediately with the
        URL the avatar will be available at once the worker completes. Clients
        can optimistically display that URL; on failure the URL will 404 and
        the user's `avatar_url` field will not be updated.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [picture]
              properties:
                picture:
                  type: string
                  format: binary
                  description: Image file (jpeg/png/webp/heic). Hard cap 50 MiB at the multipart boundary; anything between 5 MiB and 50 MiB is recompressed server-side.
      responses:
        "202":
          description: Upload accepted and queued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        required: [avatar_url, status]
                        properties:
                          avatar_url:
                            type: string
                            format: uri
                            description: URL the avatar will be available at once the background worker completes.
                          status:
                            type: string
                            enum: [processing]
              example:
                status: success
                message: Avatar upload accepted
                code: ACCEPTED
                data:
                  avatar_url: "https://cdn.fitcall.me/fitcall-avatars/avatars/550e8400-e29b-41d4-a716-446655440000/8f14e45f-ceea-467a-9575-d8d6e80b9f4a.jpg"
                  status: processing
        "400":
          description: Missing/invalid multipart, unsupported MIME, or unreadable image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: "Unsupported image format. Accepted: jpeg, png, webp, heic"
                code: BAD_REQUEST
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: File exceeds 50 MiB multipart boundary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Storage backend not configured, or upload queue is full
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                status: error
                message: Upload service is busy, please retry shortly
                code: SERVER_ERROR
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...